moved height map from towr_ros to towr
