+++ towr/include/towr/nlp_factory.h
@@ -37,7 +37,7 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #include <towr/variables/spline_holder.h>
 
 #include <towr/models/robot_model.h>
-#include "height_map.h"
+#include <towr/terrain/height_map.h>
 #include "parameters.h"
 
 namespace towr {
