+++ towr/include/towr/towr.h
@@ -38,7 +38,7 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #include <towr/variables/spline_holder.h>
 
 #include <towr/models/robot_model.h>
-#include "height_map.h"
+#include <towr/terrain/height_map.h>
 #include "nlp_factory.h"
 #include "parameters.h"
