+++ towr/src/height_map_examples.cc
@@ -27,7 +27,7 @@ OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 ******************************************************************************/
 
-#include <towr_ros/height_map_examples.h>
+#include <towr/terrain/examples/height_map_examples.h>
 
 namespace towr {
