+++ towr_ros/CMakeLists.txt
@@ -34,7 +34,6 @@ include_directories(
 ## The main executable that runs TOWR
 add_executable(towr_ros 
   src/towr_ros.cc
-  src/height_map_examples.cc
 )
 add_dependencies(towr_ros ${PROJECT_NAME}_gencpp) # build .msg first
 target_link_libraries(towr_ros
